refactor(docservice): ♻️ rename the backend to editorDataRedis and ship it ungated
editorDataRedisLocks -> editorDataRedis. This is the config value operators
already set, following upstream documentation, and setting it currently
crash-loops because no such module exists in any open-source tree.

The module ships in the standard documentserver build with no licence
gate. It stays opt-in: the locks fail closed, so selecting Redis without a
reachable Redis would refuse every save behind a healthy healthcheck.

DocService/package.json's pkg manifest has always listed
./sources/editorDataRedis.js, matching no file; that entry now resolves,
and its siblings are listed alongside it.
